Organize Imports, Merge A Couple Modules
